requirements.txt
Django==1.11.4
django-debug-toolbar==1.8
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
PyJWT==1.5.3
pytz==2017.2
sqlparse==0.2.3
